Align on first contact
Player alignment now only happens when the planet is touched for the first time after entering the gravity hitbox. One bug is that the accumulated y-rotation suddenly gets applied on firstcontact because im currently using a accumulative mouse x and y for planet rotation and delta mouse x and y for space rotation. FIX WILL BE TO CHANGE EVERYTHING TO DELTA